Trying to make a responsive system

+++ GameObjects/Bosses/FelixTheToad.cs
@@ -2,7 +2,7 @@ public class FelixTheToad : Boss
 {
     public int Steps { get; set; } = 0;
     public int SlowFrameRate { get; set; } = 0;
-    public FelixTheToad(int x, int y)
+    public FelixTheToad(float x, float y)
         // : base("Felix, the Toad", x, y, "./assets/Sprites/Bosses/Felix/F_0.pn")
         : base("Felix, the Toad", x, y, "../../../assets/Sprites/Bosses/Felix/F_0.png")
     {
